+++ packages/core/LICENSE
