+++ packages/core/fonts/open-sans/open-sans-12-black/open-sans-12-black.png
