+++ packages/core/fonts/open-sans/open-sans-128-black/open-sans-128-black.png
