+++ packages/core/fonts/open-sans/open-sans-128-white/open-sans-128-white.png
