+++ packages/core/fonts/open-sans/open-sans-16-black/open-sans-16-black.png
