+++ packages/core/fonts/open-sans/open-sans-32-black/open-sans-32-black.png
