+++ packages/core/fonts/open-sans/open-sans-32-white/open-sans-32-white.png
