+++ packages/core/fonts/open-sans/open-sans-64-black/open-sans-64-black.png
