+++ packages/core/fonts/open-sans/open-sans-64-white/open-sans-64-white.png
