+++ packages/core/fonts/open-sans/open-sans-8-black/open-sans-8-black.png
