+++ packages/core/fonts/open-sans/open-sans-8-white/open-sans-8-white.fnt
