+++ packages/core/jimp.d.ts
