+++ packages/core/karma.conf.js
