+++ packages/core/package.json
@@ -1,0 +1,52 @@
+{
+  "name": "@jimp/core",
+  "version": "0.3.5",
+  "description": "Jimp core",
+  "main": "./dist/jimp.js",
+  "types": "./jimp.d.ts",
+  "files": [
+    "dist",
+    "es",
+    "jimp.d.ts",
+    "fonts"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/oliver-moran/jimp.git"
+  },
+  "bugs": {
+    "url": "https://github.com/oliver-moran/jimp/issues"
+  },
+  "scripts": {
+    "lint": "npm run -s lint:src && npm run -s lint:tests && npm run -s lint:extra && echo Lint All Success",
+    "lint:src": "xo src/*.js src/**/*.js !src/modules/*.js",
+    "lint:tests": "xo test/*.js --global before",
+    "lint:extra": "xo {tools,examples}/*.js",
+    "test:browser": "cross-env BABEL_ENV=test ENV=browser CHROME_BIN=\"$(which chrome || which chromium)\" karma start --single-run",
+    "test:browser:watch": "npm run -s test:browser start -- --auto-watch",
+    "test:node": "cross-env BABEL_ENV=test mocha --require @babel/register",
+    "test:node:watch": "npm run test:node -- --reporter min --watch",
+    "test:node:coverage": "nyc npm run test:node",
+    "test": "npm run -s test:node:coverage && npm run -s test:browser",
+    "build": "npm run build:node:production && npm run build:module",
+    "build:debug": "npm run build:node:debug",
+    "build:module": "cross-env BABEL_ENV=module babel src -d es --source-maps",
+    "build:node": "babel src -d dist --source-maps",
+    "build:node:debug": "cross-env BABEL_ENV=development npm run build:node",
+    "build:node:production": "cross-env BABEL_ENV=production npm run build:node"
+  },
+  "author": "Oliver Moran <oliver.moran@gmail.com>",
+  "license": "MIT",
+  "dependencies": {
+    "any-base": "^1.1.0",
+    "buffer": "^5.2.0",
+    "exif-parser": "^0.1.12",
+    "file-type": "^9.0.0",
+    "load-bmfont": "^1.3.0",
+    "mkdirp": "0.5.1",
+    "phin": "^2.9.1",
+    "pixelmatch": "^4.0.2",
+    "tinycolor2": "^1.4.1"
+  },
+  "xo": false
+}
