+++ packages/core/src/image-manipulation/blur-tables.js
