+++ packages/core/src/image-manipulation/color.js
