+++ packages/core/src/image-manipulation/composite-modes.js
