+++ packages/core/src/image-manipulation/effects.js
