+++ packages/core/src/image-manipulation/shape.js
