+++ packages/core/src/image-manipulation/text.js
