+++ packages/core/src/jimp.js
