+++ packages/core/src/modules/resize.js
