+++ packages/core/src/modules/resize2.js
