+++ packages/core/src/utils/error-checking.js
