+++ packages/core/src/utils/image-bitmap.js
