+++ packages/core/src/utils/is-def.js
