+++ packages/core/src/utils/mime.js
