+++ packages/core/src/utils/scan.js
