+++ packages/core/tests/mime.test.js
